Adds alias support
This commit adds the ability for Hubot to respond to messages
directed at it via an alias. This command used to be supported
with the --enabled-slash parameter, and as a result the slash '/'
was fittingly chosen as default.

This value can easily be changed via the parameter 'hubot::chat_alias'
defined in either hiera or at class declaration.

The parameter name `chat_alias` was chosen over `alias`, as the latter
is a Puppet metaparameter.

Examples:

```yaml
---
  hubot::chat_alias: '!'
```

```ruby
  class { 'hubot':
    chat_alias => '!',
  }
```